requirements.txt
pydantic~=1.10.2
fastapi~=0.88.0
SQLAlchemy~=1.4.45
python-dotenv~=0.21.0
numpy~=1.24.1
pandas~=1.5.2
scipy~=1.9.3
torch~=1.13.1
Pillow~=9.3.0